OK, it compiles.
N-DJBDNS needs some ancient version of autoconf to compile.  To solve
this, I took the configure script and various Makefile.in files from
n-djbdns-1.06 and put them in the Git tree.  I have *also* removed
the Makefile.am and configure.ac files (they do not work with Ubuntu
20.04’s default automake, so are useless).

I believe it was a mistake to move from DJB’s build system to autoconf.

When people tried to add autoconf stuff to MaraDNS, I never let them do
it: Autoconf is a maintenance headache.  Autoconf configuration files
break from autoconf version to version; autoconf now has stupid requirements
like requiring there to be a README file (but it does not accept
a README.md file, which is the GitHub standard).